refactor(api-fragment): replace root barrel with direct modules
Move server construction and the framework-agnostic client builder into
dedicated modules. Remove the package root export, update consumers and
build entries, and publish explicit server and client subpaths.
