+++ .changeset/api-fragment-direct-modules.md
@@ -1,0 +1,5 @@
+---
+"@fragno-dev/api-fragment": patch
+---
+
+refactor: replace the package root barrel with direct server and client imports.
