+++ packages/api-fragment/src/client/react.ts
@@ -1,7 +1,7 @@
 import type { FragnoPublicClientConfig } from "@fragno-dev/core/client";
 import { createFragnoReactClient } from "@fragno-dev/core/react";
 
-import { createApiFragmentClients } from "..";
+import { createApiFragmentClients } from "./client";
 
 export function createApiFragmentClient(config: FragnoPublicClientConfig = {}) {
   return createFragnoReactClient(createApiFragmentClients(config));
