+++ packages/api-fragment/src/client/solid.ts
@@ -1,7 +1,7 @@
 import type { FragnoPublicClientConfig } from "@fragno-dev/core/client";
 import { useFragno } from "@fragno-dev/core/solid";
 
-import { createApiFragmentClients } from "..";
+import { createApiFragmentClients } from "./client";
 
 export function createApiFragmentClient(config: FragnoPublicClientConfig = {}) {
   return useFragno(createApiFragmentClients(config));
