+++ packages/api-fragment/src/client/vanilla.ts
@@ -1,7 +1,7 @@
 import type { FragnoPublicClientConfig } from "@fragno-dev/core/client";
 import { useFragno } from "@fragno-dev/core/vanilla";
 
-import { createApiFragmentClients } from "..";
+import { createApiFragmentClients } from "./client";
 
 export function createApiFragmentClient(config: FragnoPublicClientConfig = {}) {
   return useFragno(createApiFragmentClients(config));
