+++ packages/api-fragment/src/client/vue.ts
@@ -1,7 +1,7 @@
 import type { FragnoPublicClientConfig } from "@fragno-dev/core/client";
 import { useFragno } from "@fragno-dev/core/vue";
 
-import { createApiFragmentClients } from "..";
+import { createApiFragmentClients } from "./client";
 
 export function createApiFragmentClient(config: FragnoPublicClientConfig = {}) {
   return useFragno(createApiFragmentClients(config));
